Include Logback as a test dependency.
Since these libraries do logging, they need some kind of logging
framework present whenever they are run. Test dependencies are a good
way to handle this kind of requirement, especially since this project
should really have some unit tests. (Although a lot of the code in this
project is too simple and difficult to test to be worth the effort.)