Don't inline `LexerUtil`'s `chars` method.
The compiler doesn't seem to allow inlining this method--possibly
because it's defined in a trait. This method is probably a little
complex for inlining anyway, so let's just let it be a normal method
call. Building the parser combinator is a relatively infrequent
operation anyway since they're expensive enough to require reuse.