New methods for finding longest token.
These are a set of functions called `longest` or `longestString` that
take in a collection of possible constant tokens and create a lexer that
will match the longest of the group. These are designed to be as cheap
as possible at match time, so they are best reused.

The Scala parser library has a longest-match combinator (`|||`) but
it is designed for dynamic matching and must exhaustively try every
candidate. **These functions only operate on tokens that can be
represented as constant strings.**