Eliminate `quarter` method; keep `getQuarter`.
Although `quarter` is more Scala-like, it's actually inconsistent with
some other Joda methods, which give back complex objects in response
to methods with names like `year`. Trying to have one inline method
on a value class wrap another inlined method was also too much for
the inliner to handle, which means that the only way to have both be
inlineable would be to use copy-paste.