Skip indirection when constructing exception.
By default, Scala has a type alias under the name
`scala.UnsupportedOperationException`, which is simply a pointer to
`java.lang.UnsupportedOperationException`. This is there to help
decouple Scala's standard libraries from the JVM, but it just slows down
the compiler in our case--where we're generating code in a macro anyway.