Fix type signature of getCaseObjects for 2.11.
Scala 2.11's type inference seems to be a little smarter in a way
that's not so helpful to us when we're building sets. Because sets
are parametrically invariant, we need to explicitly pass in the
parametric type we want to `Set.apply`, or else we get a signature like
`Set[Product with X]` that's not compatible with the `Set[X]` that we've
declared. (The compiler's choice is correctly a narrower unification,
but it's not what we want.)