Add detection for millisecond-precision times.
When showing times, we don't want to show `3.00 ms` if the original
data was only measured at millisecond precision. We can pretty reliably
detect millisecond-precision measurements and change the format so we
only say `3 ms` in that case.

(In the JVM, you're almost always going to have either nanosecond
precision or millisecond precision for your time measurements. It's
possible that you'll have second precision in some unusual cases, but
it's not worth optimizing for them.)