New Java preferences utilities
This provides a convenient Scala-style typed mapping to and from
java.util.prefs preferences. These are mostly useful for client GUI
applications (e.g., JavaFX), though they could act as a kind of local
storage for a CLI application as well.

This also introduces the `Homomorphism` class, which is a surprising
gap in the `scalaz` library. There's not even a `Read` (aka `Parse`)
which is the adjoint of `Show`. `Show` plus `Read` is a homomorphism
with `String`.