ZipStream allows you to stream out a zip archive
This is a new Akka Streams provider that allows you to send in a bunch
of streaming zip entries and write the output to either a zip archive on
disk or receive source of streaming bytes suitable for sending over the
network.

IO streaming adds a bit of CPU overhead, but it has the advantage of
being fully nonblocking, backpressure sensitive, and free from large
buffers. All this means that it will scale extremely well horizontally.