Add "Throwing" implementations of JDBC classes
This makes it easy to get started if you're building a toy JDBC driver
and don't want hundreds of errors about unimplemented methods. This sets
things up so that every single method throws a message saying that an
unimplemented method was called.

This is really for toy implementations or as a starting place to a real
implementation. I wouldn't recommend shipping a real driver that derives
from one of these.