Use `toIterator` instead of `toStream`
We want to lazily operate on the `TraversableOnce`, but Scala's dubious
immutable `Stream` has a bad tendency to blow out the stack and generate
a lot of garbage. Using `toIterator` accomplishes our goal just as well.

Cats has even better traversal options, but I don't want to force a
dependency just for this.